\name{IrisViewer-package}
\alias{IrisViewer-package}
\alias{IrisViewer}
\docType{package}
\title{
\packageTitle{IrisViewer}
}
\description{
\packageDescription{IrisViewer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{IrisViewer}
\packageIndices{IrisViewer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{IrisViewer}

Maintainer: \packageMaintainer{IrisViewer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
